+++ adapter-auth/src/main/java/dev/jpje/jobtracker/auth/adapter/AuthenticationManagerAdapter.java
@@ -1,0 +1,31 @@
+package dev.jpje.jobtracker.auth.adapter;
+
+import java.util.Objects;
+
+import dev.jpje.jobtracker.auth.userdetails.AuthUserDetails;
+import dev.jpje.jobtracker.domain.model.User;
+import dev.jpje.jobtracker.domain.port.outbound.AuthenticateUserPort;
+import org.springframework.security.authentication.AuthenticationManager;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.AuthenticationException;
+import org.springframework.stereotype.Component;
+
+@Component
+public class AuthenticationManagerAdapter implements AuthenticateUserPort {
+  private final AuthenticationManager authenticationManager;
+
+  public AuthenticationManagerAdapter(final AuthenticationManager authenticationManager) {
+    this.authenticationManager = authenticationManager;
+  }
+
+  @Override
+  public User authenticate(final String username, final String rawPassword) {
+    try {
+      final var authentication = authenticationManager.authenticate(
+        new UsernamePasswordAuthenticationToken(username, rawPassword));
+      return ((AuthUserDetails) Objects.requireNonNull(authentication.getPrincipal())).user();
+    } catch (final AuthenticationException e) {
+      throw new IllegalArgumentException("Invalid credentials", e);
+    }
+  }
+}